--- CART/CARTLib/examples/shared/__init__.py
@@ -1,1 +1,0 @@
-from .CARTStandardUnit import CARTStandardUnit
